FACES-3235 Consolidate non-PrimeFaces issue portlets into a single issues war
