--- issue/FACES-1478-portlet/src/main/resources/log4j.properties
@@ -1,8 +1,0 @@
-log4j.rootLogger=INFO, CONSOLE
-
-log4j.appender.CONSOLE=org.apache.log4j.ConsoleAppender
-log4j.appender.CONSOLE.layout=org.apache.log4j.PatternLayout
-log4j.appender.CONSOLE.layout.ConversionPattern=%d{ABSOLUTE} %-5p [%c{1}:%L] %m%n
-
-log4j.logger.com.liferay.faces.util.lifecycle.DebugPhaseListener=DEBUG
-log4j.logger.com.liferay.faces.issues=DEBUG
