--- issue/JSFSPEC-1433-portlet/src/main/webapp/WEB-INF/faces-config.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<faces-config version="1.2" xmlns="http://java.sun.com/xml/ns/javaee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-facesconfig_1_2.xsd">
-	<application>
-		<view-handler>com.liferay.faces.bridge.application.PortletFaceletViewHandler</view-handler>
-	</application>
-	<lifecycle>
-		<phase-listener>com.liferay.faces.util.lifecycle.DebugPhaseListener</phase-listener>
-	</lifecycle>
-	<managed-bean>
-		<managed-bean-name>testBean</managed-bean-name>
-		<managed-bean-class>com.liferay.faces.issues.bean.TestBean</managed-bean-class>
-		<managed-bean-scope>request</managed-bean-scope>
-	</managed-bean>
-</faces-config>
