--- issue/JSFSPEC-1433-portlet/src/main/webapp/WEB-INF/liferay-display.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE display PUBLIC "-//Liferay//DTD Display 6.2.0//EN" "http://www.liferay.com/dtd/liferay-display_6_2_0.dtd">
-
-<display>
-	<category name="category.sample">
-		<portlet id="1" />
-	</category>
-</display>
