--- issue/JSFSPEC-1433-portlet/src/main/webapp/WEB-INF/liferay-plugin-package.properties
@@ -1,9 +1,0 @@
-name=FACES-1433
-module-group-id=com.liferay.faces.issue
-module-incremental-version=1
-tags=
-short-description=
-change-log=
-page-url=http://www.liferay.com
-author=Liferay, Inc.
-licenses=Apache License, Version 2.0
