--- issue/JSFSPEC-1433-portlet/src/main/webapp/WEB-INF/web.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0"?>
-
-<web-app xmlns="http://java.sun.com/xml/ns/javaee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_2_5.xsd" version="2.5">
-	<!-- JAVASERVERFACES-3031 -->
-	<!--
-	<context-param>
-		<param-name>com.sun.faces.namespaceParameters</param-name>
-		<param-value>true</param-value>
-	</context-param>
-	-->
-	<context-param>
-		<param-name>javax.faces.ALWAYS_PERFORM_VALIDATION_WHEN_REQUIRED_IS_TRUE</param-name>
-		<param-value>true</param-value>
-	</context-param>
-	<!-- Setting the JSF 2 PROJECT_STAGE to Development will cause the JSF implementation to do the following at runtime: -->
-	<!-- 1. Log more verbose messages. -->
-	<!-- 2. Render tips and/or warnings in the view markup. -->
-	<!-- 3. Cause the default ExceptionHandler to display a developer-friendly error page. -->
-	<context-param>
-		<param-name>javax.faces.PROJECT_STAGE</param-name>
-		<param-value>${project.stage}</param-value>
-	</context-param>
-	<servlet>
-		<servlet-name>Faces Servlet</servlet-name>
-		<servlet-class>javax.faces.webapp.FacesServlet</servlet-class>
-		<load-on-startup>1</load-on-startup>
-	</servlet>
-</web-app>
