--- issue/JSFSPEC-1433-portlet/src/main/webapp/WEB-INF/weblogic.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0"?>
-
-<wls:weblogic-web-app
-	xmlns:wls="http://xmlns.oracle.com/weblogic/weblogic-web-app"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_2_5.xsd http://xmlns.oracle.com/weblogic/weblogic-web-app http://xmlns.oracle.com/weblogic/weblogic-web-app/1.4/weblogic-web-app.xsd"
->
-	<wls:container-descriptor>
-		<wls:prefer-application-packages>
-			<wls:package-name>com.sun.xml.bind.*</wls:package-name>
-			<wls:package-name>org.apache.commons.fileupload.*</wls:package-name>
-			<wls:package-name>org.apache.xerces.*</wls:package-name>
-		</wls:prefer-application-packages>
-		<wls:optimistic-serialization>true</wls:optimistic-serialization>
-		<wls:show-archived-real-path-enabled>true</wls:show-archived-real-path-enabled>
-	</wls:container-descriptor>
-</wls:weblogic-web-app>
