+++ issue/jsf-issues-portlet/src/main/resources/log4j.properties
@@ -1,0 +1,21 @@
+log4j.rootLogger=INFO, CONSOLE
+
+log4j.appender.CONSOLE=org.apache.log4j.ConsoleAppender
+log4j.appender.CONSOLE.layout=org.apache.log4j.PatternLayout
+log4j.appender.CONSOLE.layout.ConversionPattern=%d{ABSOLUTE} %-5p [%c{1}:%L] %m%n
+
+log4j.logger.com.liferay.faces.bridge.application.internal.ApplicationImpl=DEBUG
+
+log4j.logger.com.liferay.faces.bridge.bean.internal.BeanManagerImpl=DEBUG
+log4j.logger.com.liferay.faces.bridge.bean.internal.BeanManagerMojarraImpl=DEBUG
+log4j.logger.com.liferay.faces.bridge.bean.internal.MojarraInjectionProvider=DEBUG
+log4j.logger.com.liferay.faces.bridge.bean.internal.PreDestroyInvokerMojarraImpl=DEBUG
+
+log4j.logger.com.liferay.faces.bridge.renderkit.html_basic.internal.HeadManagedBean=DEBUG
+log4j.logger.com.liferay.faces.bridge.renderkit.html_basic.internal.HeadPhaseListener=DEBUG
+log4j.logger.com.liferay.faces.bridge.renderkit.html_basic.internal.HeadRendererBridgeImpl=DEBUG
+
+log4j.logger.com.liferay.faces.bridge.servlet.BridgeSessionListener=DEBUG
+
+log4j.logger.com.liferay.faces.util.lifecycle.DebugPhaseListener=DEBUG
+log4j.logger.com.liferay.faces.issues=DEBUG
