+++ issue/jsf-issues-portlet/src/main/webapp/WEB-INF/liferay-display.xml
@@ -3,6 +3,7 @@
 
 <display>
 	<category name="category.sample">
-		<portlet id="1" />
+		<portlet id="FACES-1478" />
+		<portlet id="JSFSPEC-1433" />
 	</category>
 </display>
