+++ issue/jsf-issues-portlet/src/main/webapp/WEB-INF/liferay-plugin-package.properties
@@ -1,4 +1,4 @@
-name=FACES-1478
+name=jsf-issues
 module-group-id=com.liferay.faces.issue
 module-incremental-version=1
 tags=
