+++ issue/jsf-issues-portlet/src/main/webapp/WEB-INF/portlet.xml
@@ -3,12 +3,12 @@
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://java.sun.com/xml/ns/portlet/portlet-app_2_0.xsd http://java.sun.com/xml/ns/portlet/portlet-app_2_0.xsd">
 	<portlet>
-		<portlet-name>1</portlet-name>
+		<portlet-name>FACES-1478</portlet-name>
 		<display-name>FACES-1478</display-name>
 		<portlet-class>javax.portlet.faces.GenericFacesPortlet</portlet-class>
 		<init-param>
 			<name>javax.portlet.faces.defaultViewId.view</name>
-			<value>/WEB-INF/views/view1.xhtml</value>
+			<value>/WEB-INF/views/FACES-1478.xhtml</value>
 		</init-param>
 		<expiration-cache>0</expiration-cache>
 		<supports>
@@ -18,6 +18,36 @@
 			<title>FACES-1478</title>
 			<short-title>FACES-1478</short-title>
 			<keywords>FACES-1478</keywords>
+		</portlet-info>
+		<security-role-ref>
+			<role-name>administrator</role-name>
+		</security-role-ref>
+		<security-role-ref>
+			<role-name>guest</role-name>
+		</security-role-ref>
+		<security-role-ref>
+			<role-name>power-user</role-name>
+		</security-role-ref>
+		<security-role-ref>
+			<role-name>user</role-name>
+		</security-role-ref>
+	</portlet>
+	<portlet>
+		<portlet-name>JSFSPEC-1433</portlet-name>
+		<display-name>JSFSPEC-1433</display-name>
+		<portlet-class>javax.portlet.faces.GenericFacesPortlet</portlet-class>
+		<init-param>
+			<name>javax.portlet.faces.defaultViewId.view</name>
+			<value>/WEB-INF/views/JSFSPEC-1433.xhtml</value>
+		</init-param>
+		<expiration-cache>0</expiration-cache>
+		<supports>
+			<mime-type>text/html</mime-type>
+		</supports>
+		<portlet-info>
+			<title>JSFSPEC-1433</title>
+			<short-title>JSFSPEC-1433</short-title>
+			<keywords>JSFSPEC-1433</keywords>
 		</portlet-info>
 		<security-role-ref>
 			<role-name>administrator</role-name>
