+++ issue/jsf-issues-portlet/src/main/webapp/WEB-INF/views/FACES-1478.xhtml
@@ -8,8 +8,8 @@
 			<hr />
 			<h:outputText value="The value of the following URL:" />
 			<br />
-			<pre><h:outputText id="url" value="#{testBean.renderUrl}" /></pre>
-			<h:outputLink value="#{testBean.renderUrl}">
+			<pre><h:outputText id="url" value="#{FACES_1478Bean.renderUrl}" /></pre>
+			<h:outputLink value="#{FACES_1478Bean.renderUrl}">
 				<h:outputText value="should be the same as the href attribute of this link" />
 			</h:outputLink>
 			<br />
